Refactor: 동네 관련 내용 domain 분리
