+++ src/main/java/homeat/backend/domain/address/dto/AddressResponse.java
@@ -1,4 +1,4 @@
-package homeat.backend.domain.user.dto;
+package homeat.backend.domain.address.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.math.BigInteger;
+import java.util.List;
 
 public class AddressResponse {
 
@@ -17,5 +18,15 @@ public class AddressResponse {
         BigInteger addressId;
         String fullNm;
         String emdNm;
+    }
+
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class GetNeighborhoodResultDTO {
+        Long totalColumnCount;
+        Long totlaPageNum;
+        List<NeighborhoodResultDTO> neighborhoods;
     }
 }
